api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.5.0"
description = ""
authors = ["ArielMAJ <ariel.maj@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
requests = "^2.31.0"
loguru = "^0.7.2"
fastapi-cache2 = "^0.2.1"


[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.6.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
